pyproject.toml
[project]
name = "adversarial-tournament"
version = "0.1.0"
description = "Multi-agent adversarial prompting tournament system using Pydantic AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai>=0.0.49",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "ujson>=5.10.0",
]

[project.scripts]
adversarial-tournament = "adversarial_tournament.main:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (require API calls)",
]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/adversarial_tournament"]
